lesson17

+++ src/lesson17/Monitor.java
@@ -2,10 +2,10 @@ package lesson17;
 
 public class Monitor implements Comparable <Monitor>  {
     private String name;
-    private int screenSize;
+    private double screenSize;
 
     @Override
     public int compareTo(Monitor o) {
-        return this.screenSize - o.screenSize;
+        return (int) (this.screenSize - o.screenSize);
     }
 }
